worker/requirements.txt
playwright==1.44.0
supabase==2.0.5
python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.5
cryptography==41.0.7
aiofiles==23.2.1